+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Changelog
 * Applying template with on demand restricting will enabled on demand restricting setting ([issue](/../../issues/1727))
 * Fixed *inet*, *media* and *sdcard* always on demand restricting ([issue](/../../issues/1722))
 * Do not import version number ([issue](/../../issues/1721))
+* Display *Default* as first template name ([issue](/../../issues/1723))
+* Updated Dutch translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)